+++ examples/apollo-federation/index.ts
@@ -1,5 +1,5 @@
 import "reflect-metadata";
-import { ApolloGateway } from "@apollo/gateway";
+import { ApolloGateway, IntrospectAndCompose } from "@apollo/gateway";
 import { ApolloServer } from "apollo-server";
 
 import * as accounts from "./accounts";
@@ -8,15 +8,15 @@ import * as products from "./products";
 import * as inventory from "./inventory";
 
 async function bootstrap() {
-  const serviceList = [
-    { name: "accounts", url: await accounts.listen(3001) },
-    { name: "reviews", url: await reviews.listen(3002) },
-    { name: "products", url: await products.listen(3003) },
-    { name: "inventory", url: await inventory.listen(3004) },
-  ];
-
   const gateway = new ApolloGateway({
-    serviceList,
+    supergraphSdl: new IntrospectAndCompose({
+      subgraphs: [
+        { name: "accounts", url: await accounts.listen(3001) },
+        { name: "reviews", url: await reviews.listen(3002) },
+        { name: "products", url: await products.listen(3003) },
+        { name: "inventory", url: await inventory.listen(3004) },
+      ],
+    }),
   });
 
   const { schema, executor } = await gateway.load();