+++ src/resolvers/convert-args.ts
@@ -42,7 +42,12 @@ function generateInstanceTransformationTree(target: TypeValue): TransformationTr
     while (superClass.prototype !== undefined) {
       const superInputType = getInputType(superClass);
       if (superInputType) {
-        inputFields = [...inputFields, ...superInputType.fields!];
+        // support overwriting fields of extended types
+        const existingFieldNames = new Set(inputFields.map(field => field.name));
+        const superFields = superInputType.fields!.filter(
+          field => !existingFieldNames.has(field.name),
+        );
+        inputFields = [...inputFields, ...superFields];
       }
       superClass = Object.getPrototypeOf(superClass);
     }